Removed un needed $products array line...
Not 100% sure why the line for building a $products[] array was left in the upgrader class, it must have been some type of possibly needed array for later in the function, but was causing a "shift null" javascript error during upgrade. This line is not needed in the upgrader class and has been removed to clean up and repair the upgrader.